requirements.txt
mediapipe==0.10.14
numpy==1.24.4
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
PyQt5==5.15.11
PyQt5_sip==12.15.0
scipy==1.9.1
